+++ ui/src/com/kotcrab/vis/ui/widget/file/internal/AbstractSuggestionPopup.java
@@ -16,16 +16,12 @@
 
 package com.kotcrab.vis.ui.widget.file.internal;
 
-import com.badlogic.gdx.math.Vector2;
-import com.badlogic.gdx.scenes.scene2d.Stage;
 import com.kotcrab.vis.ui.widget.MenuItem;
 import com.kotcrab.vis.ui.widget.PopupMenu;
-import com.kotcrab.vis.ui.widget.VisTextField;
 import com.kotcrab.vis.ui.widget.file.FileChooser;
 
 /** @author Kotcrab */
 public class AbstractSuggestionPopup extends PopupMenu {
-	private static final Vector2 tmpVector = new Vector2();
 	public static final int MAX_SUGGESTIONS = 10;
 
 	final FileChooser chooser;
@@ -41,16 +37,5 @@ public class AbstractSuggestionPopup extends PopupMenu {
 		item.getShortcutCell().space(0).pad(0);
 		item.getSubMenuIconCell().size(0).space(0).pad(0);
 		return item;
-	}
-
-	protected void showMenu (Stage stage, VisTextField pathField) {
-		Vector2 pos = pathField.localToStageCoordinates(tmpVector.setZero());
-		float menuY;
-		if (pos.y - getHeight() <= 0) {
-			menuY = pos.y + pathField.getHeight() + getHeight() - 1;
-		} else {
-			menuY = pos.y + 1;
-		}
-		showMenu(stage, pos.x, menuY);
 	}
 }